+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Update styles on `LocationInput`.
 - Feature: Add `k-u-color-valid` utility.
 - Feature: Add `valid` color in `Text` component.
 - Feature: Add `without-line-height` option to `k-typographyFontSize` method.